% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{g_slr_list}
\alias{g_slr_list}
\title{Simple Logistic Regression objective and gradient}
\usage{
g_slr_list(x, y, w)
}
\arguments{
\item{x}{NxP matrix of independent variables. First column should be 1 - for intercept}

\item{y}{Vector of observations. 0 or 1.}

\item{w}{Vector of model weight. This parameter will be optimised by NLopt.}
}
\value{
List of functions: objective and gradient
}
\description{
This is a helper function that defines SLR's objective and gradient in order
to fit the parameters via NLopt.
}
