% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{logNpexp}
\alias{logNpexp}
\title{Numerically stable log(n + exp(x))}
\usage{
logNpexp(n, x)
}
\arguments{
\item{n}{Number}

\item{x}{Number}
}
\value{
log(1/(n + exp(-x)))
}
\description{
Numerically stable log(n + exp(x))
}
