% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{logsig}
\alias{logsig}
\title{Numerically stable log(plogis(x)) := -log1p(exp(-x))}
\usage{
logsig(x)
}
\arguments{
\item{x}{Vector of numbers}
}
\value{
Vector of results
}
\description{
This is a helper function to make logistic regression loss numerically stable.
Based on [this blog](https://fa.bianp.net/blog/2019/evaluate_logistic/) by Fabian Pedregosa
and [this note](https://cran.r-project.org/web/packages/Rmpfr/vignettes/log1mexp-note.pdf)
by Martin Mächler
}
