% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{mlr}
\alias{mlr}
\title{Simple Logistic Regression}
\usage{
mlr(x, s, ret_c = FALSE)
}
\arguments{
\item{x}{NxP matrix-compatible table with independent variables.}

\item{s}{Vector of binary ([0, 1]) class assignments with 1 corresponding to positive-unlabeled and 0 to unlabelled observations.}

\item{ret_c}{Logical. Whether or not to return the estimated c - probability of positive sample to be unlabeled}
}
\value{
Named vector of coefficients. If `ret_c = TRUE` first element is `c_hat`.
}
\description{
Implementation based on Jaskie et al., 2019.
}
\details{
Jaskie, Kristen, Charles Elkan, and Andreas Spanias. "A modified logistic regression for positive and unlabeled learning." 2019 53rd Asilomar Conference on Signals, Systems, and Computers. IEEE, 2019.
}
