% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{sigm_b}
\alias{sigm_b}
\title{Numerically stable `plogis(t) - y`}
\usage{
sigm_b(t, y)
}
\arguments{
\item{t}{Vector of inputs to the logistic/sigmoid function}

\item{y}{Dependent variable: [0, 1]. Should be the same length as t.}
}
\value{
Vector of outputs
}
\description{
This is a helper function to make the gradient of the logistic regression loss
numerically stable.
Based on [this blog](https://fa.bianp.net/blog/2019/evaluate_logistic/) by Fabian Pedregosa.
}
