% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{slr}
\alias{slr}
\title{Simple Logistic Regression}
\usage{
slr(x, y, print_level = NULL)
}
\arguments{
\item{x}{NxP matrix-compatible table with independent variables.}

\item{y}{Vector of binary ([0, 1]) class assignments.}
}
\value{
Vector of coefficients after fitting the model.
}
\description{
Simple Logistic Regression
}
